refactorizacion de memoria, mods en cpu varias

+++ cpu/src/iniciar_cpu.h
@@ -12,7 +12,7 @@ void finalizar_cpu();
 
 //void loggear_ejecucion(t_instruccion *instruccion);//log obligatorio
 int get_tamanio_registro(char *registro);
-long int get_registro(char *registro);
+uint32_t get_registro(char *registro);
 int consultar_tamanio_pagina_memoria();
 
 #endif
